# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development — بيئة التطوير"
    },
    {
      "url": "https://akfeek-backend.developteam.site",
      "description": "Production — بيئة الإنتاج"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token — أدخل التوكن بعد كلمة Bearer مثال: `Bearer eyJhbGci...`"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          },
          "errorAr": {
            "type": "string",
            "example": "رسالة الخطأ"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "مطلوب مصادقة | Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "error": "Authentication required",
              "errorAr": "مطلوب مصادقة",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "تم رفض الوصول | Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "error": "Access denied",
              "errorAr": "تم رفض الوصول",
              "code": "FORBIDDEN"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "🚗 Akfeek API — Full Reference",
    "version": "1.0.0",
    "description": "\n# Akfeek Platform — Full API Reference\n\n> **للمطورين الداخليين فقط** — هذه الوثيقة تحتوي على **جميع** الـ endpoints.\n> لو بتشتغل على جزء معين، استخدم الرابط المخصص ليك:\n\n| الجمهور | الرابط |\n|---|---|\n| 📱 تطبيق موبايل العميل | [`/api-docs/mobile`](/api-docs/mobile) |\n| 🔧 تطبيق موبايل الفني | [`/api-docs/technician`](/api-docs/technician) |\n| 🏪 بوابة الفيندور / الورشة | [`/api-docs/vendor`](/api-docs/vendor) |\n| ⚙️ لوحة تحكم الأدمن | [`/api-docs/admin`](/api-docs/admin) |\n\n---\n\n## Authentication / المصادقة\n\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n1. `POST /api/auth/login` — سجل دخول واحصل على التوكن\n2. انقر **Authorize** في الأعلى وأدخل: `Bearer <token>`\n      ",
    "contact": {
      "name": "Akfeek API Support",
      "email": "support@akfeek.com"
    }
  },
...
  "paths": {
...
    "/api/brands": {
      "get": {
        "summary": "Get all vehicle brands",
        "description": "Retrieve a list of all vehicle brands (Toyota, BMW, etc.) with optional filters - ???? ??? ????? ????? ?????? ????????",
        "tags": [
          "🔓 Brands & Models"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "includeModels",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include models for each brand - ????? ??????? ?? ?????"
          },
          {
            "in": "query",
            "name": "activeOnly",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Only return active brands - ??? ???????? ?????? ???"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved brands - ?? ??? ???????? ?????",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "uuid-here"
                          },
                          "name": {
                            "type": "string",
                            "example": "Toyota"
                          },
                          "nameAr": {
                            "type": "string",
                            "example": "??????"
                          },
                          "logo": {
                            "type": "string",
                            "nullable": true,
                            "example": "https://example.com/toyota-logo.png"
                          },
                          "isActive": {
                            "type": "boolean",
                            "example": true
                          },
                          "_count": {
                            "type": "object",
                            "properties": {
                              "models": {
                                "type": "integer",
                                "example": 5
                              }
                            }
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 10
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new brand (Admin only)",
        "description": "Create a new vehicle brand - ????? ????? ????? ?????",
        "tags": [
          "⚙️ Admin | Brands & Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Tesla",
                    "description": "Brand name - ??? ???????"
                  },
                  "nameAr": {
                    "type": "string",
                    "example": "????",
                    "description": "Arabic brand name - ??? ??????? ????????"
                  },
                  "logo": {
                    "type": "string",
                    "nullable": true,
                    "example": "https://example.com/tesla-logo.png",
                    "description": "Brand logo URL - ???? ???? ???????"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Brand created successfully - ?? ????? ??????? ?????",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string",
                      "example": "Brand created successfully"
                    },
                    "messageAr": {
                      "type": "string",
                      "example": "?? ????? ??????? ?????"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error - ??? ?? ??????"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "409": {
            "description": "Brand already exists - ??????? ?????? ??????"
          }
        }
      }
    },
    "/api/brands/{id}": {
      "get": {
        "summary": "Get brand by ID",
        "description": "Retrieve a specific brand with all its models - ???? ??? ????? ????? ?? ???? ?????????",
        "tags": [
          "🔓 Brands & Models"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Brand ID - ???? ???????"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand found - ?? ?????? ??? ???????"
          },
          "404": {
            "description": "Brand not found - ??????? ??? ??????",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Brand not found"
                    },
                    "errorAr": {
                      "type": "string",
                      "example": "??????? ??? ??????"
                    },
                    "code": {
                      "type": "string",
                      "example": "BRAND_NOT_FOUND"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update brand (Admin only)",
        "description": "Update an existing brand - ????? ????? ??????",
        "tags": [
          "⚙️ Admin | Brands & Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Brand ID - ???? ???????"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "nameAr": {
                    "type": "string"
                  },
                  "logo": {
                    "type": "string",
                    "nullable": true
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Brand updated successfully - ?? ????? ??????? ?????"
          },
          "404": {
            "description": "Brand not found - ??????? ??? ??????"
          },
          "409": {
            "description": "Brand name already exists - ??? ??????? ????? ??????"
          }
        }
      },
      "delete": {
        "summary": "Delete brand (Admin only)",
        "description": "Delete a brand (soft delete by default) - ??? ????? (??? ???????)",
        "tags": [
          "⚙️ Admin | Brands & Models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Brand ID - ???? ???????"
          },
          {
            "in": "query",
            "name": "hardDelete",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Permanently delete brand and all models - ??? ????? ??????? ????? ?????????"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand deleted/deactivated - ?? ???/????? ???????"
          },
          "404": {
            "description": "Brand not found - ??????? ??? ??????"
          }
        }
      }
    },
...
    "/api/vehicles/brands": {
      "get": {
        "summary": "Get vehicle brands catalog",
        "description": "Get available vehicle brands (Toyota, BMW, Mercedes, etc.)\n\nالحصول على العلامات التجارية للمركبات المتاحة\n",
        "tags": [
          "📱 Customer | Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle brands list"
          }
        }
      }
    },
...
  }
}